# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SiteSort API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/members:
    get:
      operationId: listProjectMembers
      tags: [team]
      summary: List members of a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectMember"
    post:
      operationId: addProjectMember
      tags: [team]
      summary: Add a member to a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddMemberRequest"
      responses:
        "201":
          description: Member added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectMember"
  /projects/{projectId}/members/{memberId}:
    delete:
      operationId: removeProjectMember
      tags: [team]
      summary: Remove a member from a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: memberId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Member removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /subcontractors:
    get:
      operationId: listSubcontractors
      tags: [subcontractors]
      summary: List all subcontractors for the company
      responses:
        "200":
          description: List of subcontractors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subcontractor"
    post:
      operationId: createSubcontractor
      tags: [subcontractors]
      summary: Add a new subcontractor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubcontractorRequest"
      responses:
        "201":
          description: Subcontractor created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subcontractor"
  /subcontractors/{subcontractorId}:
    get:
      operationId: getSubcontractor
      tags: [subcontractors]
      summary: Get a subcontractor by ID
      parameters:
        - name: subcontractorId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Subcontractor details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubcontractorDetail"
    patch:
      operationId: updateSubcontractor
      tags: [subcontractors]
      summary: Update a subcontractor
      parameters:
        - name: subcontractorId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSubcontractorRequest"
      responses:
        "200":
          description: Updated subcontractor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subcontractor"
  /subcontractors/{subcontractorId}/insurance:
    post:
      operationId: addInsuranceRecord
      tags: [insurance]
      summary: Add an insurance record for a subcontractor
      parameters:
        - name: subcontractorId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddInsuranceRequest"
      responses:
        "201":
          description: Insurance record added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InsuranceRecord"
...
  /projects/{projectId}/qr-codes:
    post:
      operationId: generateQrCodes
      tags: [qr]
      summary: Generate QR codes for a project's site board
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateQrRequest"
      responses:
        "200":
          description: QR codes generated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QrCode"
  /qr/{token}:
    get:
      operationId: getQrContent
      tags: [qr]
      summary: Get documents accessible via a QR code (public or authenticated)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: QR code content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrContent"
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Document:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        uploadedBy:
          type: string
        uploaderName:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [drawing, method_statement, permit, safety, general]
        version:
          type: integer
        fileUrl:
          type: string
        fileSize:
          type: integer
        previousVersionId:
          type: string
          nullable: true
        status:
          type: string
          enum: [current, superseded]
        requiresAcknowledgment:
          type: boolean
        publicAccess:
          type: boolean
        createdAt:
          type: string
          format: date-time
        distributionSummary:
          $ref: "#/components/schemas/DistributionSummary"
      required: [id, projectId, uploadedBy, uploaderName, name, type, version, fileUrl, fileSize, status, requiresAcknowledgment, publicAccess, createdAt, distributionSummary]
...
    DistributionSummary:
      type: object
      properties:
        total:
          type: integer
        pending:
          type: integer
        viewed:
          type: integer
        acknowledged:
          type: integer
      required: [total, pending, viewed, acknowledged]
...
    ProjectMember:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        userId:
          type: string
          nullable: true
        subcontractorId:
          type: string
          nullable: true
        name:
          type: string
        role:
          type: string
          enum: [manager, worker, subcontractor]
        complianceStatus:
          type: string
          enum: [ok, warning, hold]
        addedAt:
          type: string
          format: date-time
      required: [id, projectId, name, role, complianceStatus, addedAt]
    AddMemberRequest:
      type: object
      properties:
        userId:
          type: string
          nullable: true
        subcontractorId:
          type: string
          nullable: true
        role:
          type: string
          enum: [manager, worker, subcontractor]
      required: [role]
    Subcontractor:
      type: object
      properties:
        id:
          type: string
        companyId:
          type: string
        companyName:
          type: string
        contactName:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
          nullable: true
        trades:
          type: array
          items:
            type: string
        reliabilityRating:
          type: number
          nullable: true
        paymentHold:
          type: boolean
        insuranceStatus:
          type: string
          enum: [valid, expiring_soon, expired, none]
        createdAt:
          type: string
          format: date-time
      required: [id, companyId, companyName, contactName, contactEmail, trades, paymentHold, insuranceStatus, createdAt]
    SubcontractorDetail:
      allOf:
        - $ref: "#/components/schemas/Subcontractor"
        - type: object
          properties:
            insuranceRecords:
              type: array
              items:
                $ref: "#/components/schemas/InsuranceRecord"
            assignedProjects:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                required: [id, name]
          required: [insuranceRecords, assignedProjects]
    CreateSubcontractorRequest:
      type: object
      properties:
        companyName:
          type: string
        contactName:
          type: string
        contactEmail:
          type: string
          format: email
        contactPhone:
          type: string
        trades:
          type: array
          items:
            type: string
      required: [companyName, contactName, contactEmail, trades]
    UpdateSubcontractorRequest:
      type: object
      properties:
        companyName:
          type: string
        contactName:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        trades:
          type: array
          items:
            type: string
        reliabilityRating:
          type: number
        paymentHold:
          type: boolean
    InsuranceRecord:
      type: object
      properties:
        id:
          type: string
        subcontractorId:
          type: string
        type:
          type: string
          enum: [public_liability, employers, professional]
        certificateUrl:
          type: string
        expiryDate:
          type: string
          format: date
        status:
          type: string
          enum: [valid, expiring_soon, expired]
        createdAt:
          type: string
          format: date-time
      required: [id, subcontractorId, type, certificateUrl, expiryDate, status, createdAt]
    AddInsuranceRequest:
      type: object
      properties:
        type:
          type: string
          enum: [public_liability, employers, professional]
        certificateUrl:
          type: string
        expiryDate:
          type: string
          format: date
      required: [type, certificateUrl, expiryDate]
...
    QrCode:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        category:
          type: string
        token:
          type: string
        qrImageUrl:
          type: string
        label:
          type: string
        requiresLogin:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, projectId, category, token, qrImageUrl, label, requiresLogin, createdAt]
    GenerateQrRequest:
      type: object
      properties:
        categories:
          type: array
          items:
            type: string
            enum: [safety, emergency, drawings, general]
      required: [categories]
    QrContent:
      type: object
      properties:
        projectName:
          type: string
        category:
          type: string
        requiresLogin:
          type: boolean
        documents:
          type: array
          items:
            $ref: "#/components/schemas/Document"
      required: [projectName, category, requiresLogin, documents]